refactor(cells): decompose cells into modular domain packages and pluggable router (#2149)
